Scene save/load round trip passes. - Added DeferredEventBuffer allowing users to enqueue actions or events until later - Used said buffer to defer the response of the scene load event until after all objects have been created so they're immediately available - Made LoadSceneEvent into a RequestEvent that returns success - Fixed bug in lua serializer for integer types above 32 bits - Moved LuaNodeTests to test projects and added some tests as part of investigating the serializer problem

+++ syx/Engine.Shared/lua/LuaNode.cpp
@@ -537,7 +537,7 @@ namespace Lua {
 
   void Uint64Node::_readFromLua(lua_State* s, void* base) const {
     static_assert(sizeof(lua_Integer) >= sizeof(uint64_t), "Lua number must be big enough to hold value or this needs to be rewritten");
-    _cast(base) = static_cast<size_t>(lua_tointeger(s, -1));
+    _cast(base) = static_cast<uint64_t>(lua_tointeger(s, -1));
   }
 
   void Uint64Node::_writeToLua(lua_State* s, const void* base) const {
